+++ NEWS.md
@@ -1,6 +1,7 @@
 # googleCloudStorageR 0.5.1.9000
 
 * Let users auth via an email again via `gcs_auth`
+* correct file.copy behaviour in `gcs_load_all()` (#112) - thanks @jasonmhoule
 
 # googleCloudStorageR 0.5.1