+++ NEWS.md
@@ -4,6 +4,7 @@
 * correct file.copy behaviour in `gcs_load_all()` (#112) - thanks @jasonmhoule
 * Allow GCS uploads if ACL has been set at bucket level (#111)
 * Let users set the boundary between simple and resumable upload limit via new option `option(googleCloudStorageR.upload_limit)` or `gcs_upload_set_limit()` - default is 5000000L or 5MB (#120)
+* Suppress warning for when resumable uploads are requesting the upload URL (#120)
 
 # googleCloudStorageR 0.5.1